delete btn and route

+++ public/delete-artwork.js
@@ -1,27 +1,66 @@
-async function deleteArtworkHandler(event) {
-  event.preventDefault();
+// async function deleteArtworkHandler(event) {
+//   event.preventDefault();
 
-  const id = window.location.toString().split('/')[
-    window.location.toString().split('/').length - 1
-  ];
+//   // const id = window.location.toString().split('/')[
+//   //   window.location.toString().split('/').length - 1
+//   // ];
+//   console.log('hello');
+//   console.log(id);
 
-  const response = await fetch(`/api/Artwork/${id}`, {
-    method: 'DELETE',
-    body: JSON.stringify({
-      artwork_id: id,
-    }),
-    headers: {
-      'Content-Type': 'application/json',
-    },
-  });
+//   // const response = await fetch(`/api/post/${id}`, {
+//   //   method: 'DELETE',
+//   //   body: JSON.stringify({
+//   //     artwork_id: id,
+//   //   }),
+//   //   headers: {
+//   //     'Content-Type': 'application/json',
+//   //   },
+//   // });
 
-  if (response.ok) {
-    document.location.replace('/profile/');
-  } else {
-    alert(response.statusText);
+//   if (response.ok) {
+//     document.location.replace('/api/users/artworks');
+//   } else {
+//     alert(response.statusText);
+//   }
+// }
+
+const delButtonHandler = async (event) => {
+  if (event.target.hasAttribute('data-id')) {
+    const id = event.target.getAttribute('data-id');
+    console.log('hello');
+    console.log(id);
+    const response = await fetch(`/api/post/${id}`, {
+      method: 'DELETE',
+    });
+
+    if (response.ok) {
+      document.location.replace('/api/users/artworks');
+    } else {
+      alert('Failed to delete project');
+    }
   }
-}
+};
 
 document
-  .querySelector('.delete-artwork-btn')
-  .addEventListener('click', deleteArtworkHandler);
+  .querySelector('.artwork-list')
+  .addEventListener('click', delButtonHandler);
+
+// const delButtonHandler = async (event) => {
+//   if (event.target.hasAttribute('data-id')) {
+//     const id = event.target.getAttribute('data-id');
+
+//     const response = await fetch(`/api/post/${id}`, {
+//       method: 'DELETE',
+//     });
+
+//     if (response.ok) {
+//       document.location.replace('/');
+//     } else {
+//       alert('Failed to delete artwork');
+//     }
+//   }
+// };
+
+// document
+//   .querySelector('#delete-btn')
+//   .addEventListener('click', delButtonHandler);
